feat(api): move saveAdminMember usecase
