+++ api/src/team/domain/usecases/index.js
@@ -4,6 +4,8 @@ import { fileURLToPath } from 'node:url';
 import * as mailService from '../../../../lib/domain/services/mail-service.js';
 import * as organizationRepository from '../../../../src/shared/infrastructure/repositories/organization-repository.js';
 import * as certificationCenterRepository from '../../../certification/shared/infrastructure/repositories/certification-center-repository.js';
+import * as userRepository from '../../../identity-access-management/infrastructure/repositories/user.repository.js';
+import { adminMemberRepository } from '../../../shared/infrastructure/repositories/admin-member.repository.js';
 import * as membershipRepository from '../../../shared/infrastructure/repositories/membership-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
@@ -17,6 +19,7 @@ import { organizationInvitationService } from '../services/organization-invitati
 const path = dirname(fileURLToPath(import.meta.url));
 
 const dependencies = {
+  adminMemberRepository,
   certificationCenterRepository,
   certificationCenterInvitationRepository,
   prescriberRepository,
@@ -27,6 +30,7 @@ const dependencies = {
   organizationInvitationService,
   organizationInvitationRepository,
   organizationRepository,
+  userRepository,
 };
 
 const usecasesWithoutInjectedDependencies = {
