+++ api/src/team/domain/usecases/save-admin-member.usecase.js
@@ -1,5 +1,5 @@
-import { AdminMember } from '../../../src/shared/domain/models/AdminMember.js';
-import { AlreadyExistingAdminMemberError } from '../../../src/team/domain/errors.js';
+import { AdminMember } from '../../../shared/domain/models/AdminMember.js';
+import { AlreadyExistingAdminMemberError } from '../errors.js';
 
 const saveAdminMember = async function ({ email, role, userRepository, adminMemberRepository }) {
   const { id: userId, firstName, lastName } = await userRepository.getByEmail(email);
